\encoding{utf8} 
\name{cohort1}
\docType{data}
\alias{cohort1}
\title{Simulated cohort data}
\description{
  This data corresponds to a simulated cohort with a follow-up of 32 years, including the annual radiation dose received by each subject.  
}
\usage{cohort1}
\format{A data frame with 1000 rows and 70 columns.}

\keyword{ datasets }