\encoding{utf8} 
\name{linERR-internal}
\alias{linERR-internal}
\alias{summary.fit.linERR}
\alias{print.summary.fit.linERR}
\alias{plot.fit.linERR}
\alias{llhood}

\title{Internal fit.linERR functions}
\description{Internal fit.linERR functions}
\usage{
\method{summary}{fit.linERR}(object, ...)
\method{print.summary}{fit.linERR}(x, ...)
\method{plot}{fit.linERR}(x, ...)
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{linERR-package}}, \code{\link{fit.linERR}}, \code{\link{ERRci}}
}
\keyword{internal}
